utils/honeypot.py
import logging
from telegram import Update
from telegram.ext import ContextTypes
import database as db

logger = logging.getLogger(__name__)

async def monitor_group(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """يراقب المجموعات بحثاً عن روابط مشبوهة إذا تم تفعيل Honeypot."""
    message = update.effective_message
    if not message or not message.text: return
    
    # منطق بسيط لاصطياد الروابط في المجموعات
    if "http" in message.text:
         # هنا يمكن إضافة منطق الفحص التلقائي لكل رابط يرسل في المجموعات المشتركة
         pass
