requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.5
apscheduler==3.10.4
feedparser==6.0.10
Pillow==10.1.0
stripe==7.14.0
openai>=1.6.0
filetype==1.2.0
